refactor: redução de consumo de memória e aumenta intensidade de luzes

+++ trabalho_final/include/commons.hpp
@@ -8,19 +8,19 @@
 
 struct Material {
     Vec3 color;
-    double kd, ks, ka;
-    double shininess;
+    float kd, ks, ka;
+    float shininess;
     bool useTexture = false;
     Texture* texture = nullptr;
 };
 
 struct HitRecord {
-    double t;         // Distancia da interseção
+    float t;         // Distancia da interseção
     Vec3 p;           // Ponto exato no espaco (x,y,z)
     Vec3 normal;      // Vetor normal
     int objectID;
     Material mat;
-    double u, v;      // coordenadas de textura  
+    float u, v;      // coordenadas de textura  
 };
 
 #endif
