+++ trabalho_final/include/cone.hpp
@@ -6,10 +6,10 @@
 class Cone : public Object {
 public:
     Vec3 cBase, V, n;
-    double raio, H, theta;
+    float raio, H, theta;
     bool base;
 
-    Cone(Vec3 base, Vec3 eixo, double r_param, double altura, Material m, bool baseOn = true, const char* nome = "") {
+    Cone(Vec3 base, Vec3 eixo, float r_param, float altura, Material m, bool baseOn = true, const char* nome = "") {
         cBase = base;
         n = eixo.normalize();
         raio = r_param;
@@ -22,12 +22,12 @@ public:
     }
 
     float intersectBase(const Ray& r) {
-        double rBase = raio;
-        double denom = r.direction.dot(n);
+        float rBase = raio;
+        float denom = r.direction.dot(n);
 
         if(std::abs(denom) < 1e-6) return -1.0f;
 
-        double t = (cBase - r.origin).dot(n) / denom;
+        float t = (cBase - r.origin).dot(n) / denom;
         if (t < 0.001) return -1.0f;
 
         Vec3 P = r.pt(t);
@@ -38,36 +38,36 @@ public:
     bool intersect(const Ray& r, HitRecord& rec) override {
         Vec3 d = r.direction;
         Vec3 v = V - r.origin;
-        double cos2 = cos(theta) * cos(theta);
+        float cos2 = cos(theta) * cos(theta);
 
-        double dn = d.dot(n);
-        double vn = v.dot(n);
-        double vd = v.dot(d);
-        double dd = d.dot(d);
-        double vv = v.dot(v);
+        float dn = d.dot(n);
+        float vn = v.dot(n);
+        float vd = v.dot(d);
+        float dd = d.dot(d);
+        float vv = v.dot(v);
 
-        double a = (dn * dn) - (dd * cos2);
-        double b = (vd * cos2) - (vn * dn);
-        double c = (vn * vn) - (vv * cos2);
+        float a = (dn * dn) - (dd * cos2);
+        float b = (vd * cos2) - (vn * dn);
+        float c = (vn * vn) - (vv * cos2);
 
         float t_hit = -1.0f;
 
         if (std::abs(a) < 1e-6) {
             if (std::abs(b) > 1e-6) t_hit = -c / (2.0 * b);
         } else {
-            double delta = b * b - a * c;
+            float delta = b * b - a * c;
             if (delta >= 0) {
-                double sqrt_delta = std::sqrt(delta);
-                double t1 = (-b - sqrt_delta)/ a;
-                double t2 = (-b + sqrt_delta)/ a;
+                float sqrt_delta = std::sqrt(delta);
+                float t1 = (-b - sqrt_delta)/ a;
+                float t2 = (-b + sqrt_delta)/ a;
 
                 if (t1 > t2) std::swap(t1, t2);
 
-                for (double t_teste: {t1, t2}){
+                for (float t_teste: {t1, t2}){
                     if (t_teste > 0.001){
                         Vec3 P = r.pt(t_teste);
 
-                        double h_check = (V - P).dot(n);
+                        float h_check = (V - P).dot(n);
                         if (h_check >= 0 && h_check <= H) {
                             t_hit = t_teste;
                             break;
@@ -77,7 +77,7 @@ public:
             }
         }
 
-        double t_base = -1.0f;
+        float t_base = -1.0f;
         if (base) {
             t_base = intersectBase(r);
         }
@@ -93,7 +93,7 @@ public:
         return false;
     }
 
-    void rotateX(double angle){
+    void rotateX(float angle){
         cBase = cBase.rotX(angle);
         V = V.rotX(angle);
         n = (V - cBase).normalize();
@@ -101,21 +101,21 @@ public:
         theta = std::atan2(raio, H);
     }
 
-    void rotateY(double angle){
+    void rotateY(float angle){
         cBase = cBase.rotY(angle);
         V = V.rotY(angle);
         n = (V - cBase).normalize();
         theta = std::atan2(raio, H);
     }
 
-    void rotateZ(double angle){
+    void rotateZ(float angle){
         cBase = cBase.rotZ(angle);
         V = V.rotZ(angle);
         n = (V - cBase).normalize();
         theta = std::atan2(raio, H);
     }
 
-    void translate( double tx, double ty, double tz){
+    void translate( float tx, float ty, float tz){
         cBase = cBase.translate(tx, ty, tz);
         V = V.translate(tx, ty, tz);
     }
@@ -124,7 +124,7 @@ public:
         translate(t.x, t.y, t.z);
     }
 
-    void scaleTransform(double sx, double sy, double sz) {
+    void scaleTransform(float sx, float sy, float sz) {
         // Transformar base e vértice
         cBase = cBase.scale(sx, sy, sz);
         V = V.scale(sx, sy, sz);
@@ -139,11 +139,11 @@ public:
         theta = std::atan2(raio, H);
     }
 
-    void scaleTransform(double s) {
+    void scaleTransform(float s) {
         scaleTransform(s, s, s);
     }
 
-    void shearTransform(double xy, double xz, double yx, double yz, double zx, double zy) {
+    void shearTransform(float xy, float xz, float yx, float yz, float zx, float zy) {
         cBase = cBase.shear(xy, xz, yx, yz, zx, zy);
         V = V.shear(xy, xz, yx, yz, zx, zy);
         
@@ -156,13 +156,13 @@ public:
 
 
 private:
-    void hitLateral(HitRecord& rec, const Ray& r, double t){
+    void hitLateral(HitRecord& rec, const Ray& r, float t){
         rec.t = t;
         rec.p = r.pt(t);
 
         Vec3 VP = rec.p - V;
-        double h = VP.dot(n);
-        double k = (raio*raio)/(H*H);
+        float h = VP.dot(n);
+        float k = (raio*raio)/(H*H);
 
         Vec3 n_dir = VP - n * (h * (1.0 + k));
         rec.normal = n_dir.normalize();
@@ -175,7 +175,7 @@ private:
         rec.objectID = id;
     }
 
-    void hitBase(HitRecord& rec, const Ray& r, double t){
+    void hitBase(HitRecord& rec, const Ray& r, float t){
         rec.t = t;
         rec.p = r.pt(t);
         rec.normal = n * -1.0;
